+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMMWCWK3Z43ZXJ79M3J2QHYK"
+        "01KMPYE81RGCFZJ8A5JMH2R08Q"
     )
     r"""The unique identifier of the data source"""